Removed inconsistent debug printout
None of the other services print "Connecting to x in y". Either this is removed or the others get a similar printout. I'll opt for the former for brevity, but perhaps there is value in printing out the region before connecting, in case connection fails and the error message doesn't provide the value of the region.