Update cloudformation template to work around pip permissions issue
Running as the root user, easy_install installs pip into
`/usr/local/bin`, which is not in the root users path. This causes the
subsequent calls to pip to fail. This change invokes pip via an absolute
path to get around that.

An additional complication due to pip being installed in
`/usr/local/bin` is that it installs dynamic-dynamodb into
`/usr/local/bin` as well. This change therefor also updates the init.d
file to reference the binary correctly.